+++ CHANGELOG.md
@@ -11,6 +11,12 @@ Changelog
 
 _Note: Gaps between patch versions are faulty, broken or test releases._
 
+## v3.0.0-rc.??? (2021-06-??)
+
+#### :bug: Bug Fix
+
+* Fixed caching of form values `bForm`
+
 ## v3.0.0-rc.202 (2021-06-18)
 
 #### :bug: Bug Fix